Added some controller tests

+++ src/main/java/br/com/cleanarchitecture/starwars/core/usecases/ManagePlanetUseCaseImpl.java
@@ -2,7 +2,10 @@ package br.com.cleanarchitecture.starwars.core.usecases;
 
 import br.com.cleanarchitecture.starwars.core.domain.Planet;
 import br.com.cleanarchitecture.starwars.core.exceptions.PlanetAlreadyExistsException;
+import br.com.cleanarchitecture.starwars.core.exceptions.PlanetNotFoundException;
 import br.com.cleanarchitecture.starwars.core.ports.repository.PlanetRepository;
+import br.com.cleanarchitecture.starwars.core.ports.thirdpartyapi.PlanetThirdPartyApi;
+import br.com.cleanarchitecture.starwars.infrastructure.thirdpartyapi.exceptions.SwapiPlanetNotFoundException;
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 
@@ -13,14 +16,20 @@ import java.util.Optional;
 public class ManagePlanetUseCaseImpl implements ManagePlanetUseCase {
     private final PlanetRepository planetRepository;
 
+    private final PlanetThirdPartyApi planetThirdPartyApi;
+
     @Override
     public Planet save(Planet planet) {
-        if (planet == null) {
+        if (planet == null || StringUtils.isBlank(planet.getName())) {
             throw new IllegalArgumentException("Invalid planet.");
         }
 
         if (findByName(planet.getName()).isPresent()) {
             throw new PlanetAlreadyExistsException();
+        }
+
+        if (!planetThirdPartyApi.isPlanetExists(planet.getName())) {
+            throw new SwapiPlanetNotFoundException();
         }
 
         return planetRepository.save(planet);
@@ -37,7 +46,12 @@ public class ManagePlanetUseCaseImpl implements ManagePlanetUseCase {
             throw new IllegalArgumentException("Invalid planet name.");
         }
 
-        return planetRepository.findOneByName(name);
+        Optional<Planet> planet = planetRepository.findOneByName(name);
+        if (!planet.isPresent()) {
+            throw new PlanetNotFoundException();
+        }
+
+        return planet;
     }
 
     @Override
@@ -46,13 +60,23 @@ public class ManagePlanetUseCaseImpl implements ManagePlanetUseCase {
             throw new IllegalArgumentException("Invalid planet id.");
         }
 
-        return planetRepository.findOneById(id);
+        Optional<Planet> planet = planetRepository.findOneById(id);
+        if (!planet.isPresent()) {
+            throw new PlanetNotFoundException();
+        }
+
+        return planet;
     }
 
     @Override
     public void delete(String id) {
         if (StringUtils.isBlank(id)) {
             throw new IllegalArgumentException("Invalid planet id.");
+        }
+
+        Optional<Planet> planet = findById(id);
+        if (!planet.isPresent()) {
+            throw new PlanetNotFoundException();
         }
 
         planetRepository.delete(id);
